Moved model mockup classes to correct package.
