+++ msom/src/main/java/pl/edu/uj/fais/wpz/msom/model/mockups/ProcessingSystemMockup.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package pl.edu.uj.fais.wpz.msom.model;
+package pl.edu.uj.fais.wpz.msom.model.mockups;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,7 +29,7 @@ import pl.edu.uj.fais.wpz.msom.service.interfaces.TaskTypeService;
  *
  * @author jarep
  */
-public class ProcessingSystemMockup extends AbstractModelObject<Model> implements ProcessingSystem {
+public class ProcessingSystemMockup extends AbstractModelMockupObject<Model> implements ProcessingSystem {
 
     private final ControllerUnitService controllerUnitService;
     private final DistributionService distributionService;
