+++ msom/src/main/java/pl/edu/uj/fais/wpz/msom/model/mockups/ProcessingUnitMockup.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package pl.edu.uj.fais.wpz.msom.model;
+package pl.edu.uj.fais.wpz.msom.model.mockups;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,7 +25,7 @@ import pl.edu.uj.fais.wpz.msom.service.interfaces.TaskTypeService;
  *
  * @author jarep
  */
-public class ProcessingUnitMockup extends AbstractModelObject<Module> implements ProcessingUnit {
+public class ProcessingUnitMockup extends AbstractModelMockupObject<Module> implements ProcessingUnit {
 
     private final ControllerUnitService controllerUnitService;
     private final ModuleService moduleService;
